+++ test/testmcdm.jl
@@ -1,5 +1,26 @@
 @testset "MCDM functions" begin
 
+    @testset "mcdm() with default method" begin 
+        tol = 0.00001
+        df = DataFrame()
+        df[:, :x] = Float64[9, 8, 7]
+        df[:, :y] = Float64[7, 7, 8]
+        df[:, :z] = Float64[6, 9, 6]
+        df[:, :q] = Float64[7, 6, 6]
+        w = Float64[4, 2, 6, 8]
+
+        fns = makeminmax([maximum, maximum, maximum, maximum])
+
+        result = topsis(df, w, fns)
+
+        setting = MCDMSetting(df, w, fns)
+        result2 = mcdm(setting)
+        
+        @test isa(result2, TopsisResult)
+        @test result2.bestIndex == result.bestIndex
+        @test result2.scores == result.scores
+    end 
+
     @testset "MOOSRA" begin 
         tol = 0.0001